+++ src/document-controller.js
@@ -19,10 +19,15 @@ class DocumentController {
          */
         this.grammar = textMateGrammar;
         /**
-         *  Stores the state for each line 
+         * Stores the state of each line 
          * @type {vsctm.ITokenizeLineResult[]} 
          */
-        this.grammarState = [];
+        this.tokensArray = [];
+        /**
+         * Store document text of each line.
+         * @type {string[]}
+         */
+        this.documentText = [];
         /**
          * Parse the whole document
          */     
@@ -34,23 +39,22 @@ class DocumentController {
          */
         this.contentChangesArray = [];  // Stores text-change
         /**
-         * Store the disposable objects.
-         * @type {vscode.Disposable[]}
+         * Store the subscriptions
          */
         this.subscriptions = [];
         /**
          * onChangeDocument
          */
         this.subscriptions.push(vscode.workspace.onDidChangeTextDocument((e) => {
-            if (e.document == this.document){
-                const changes = [...e.contentChanges].sort((change1, change2) => change1.range.start.isAfter(change2.range.start) ? -1 : 1);
+            if (e.document == this.document && e.contentChanges.length){
+                const changes = [...e.contentChanges].sort((change1, change2) => change1.range.start.isAfter(change2.range.start) ? 1 : -1);
                 this.contentChangesArray = changes;
                 this.applyChanges(changes)
             }
         }));
     }
     refresh(){
-        this.grammarState = [];
+        this.tokensArray = [];
         const docRange = new vscode.Range(0, 0, this.document.lineCount, 0)
         this.reparseRange(docRange); 
         this.contentChangesArray = [];
@@ -60,14 +64,22 @@ class DocumentController {
     }
     /** parser range */
     /**
-     * 
      * @param {vscode.TextLine} line 
      */
     reparseLine(line){
-        if(!this.grammar) return;
-        const prevState = line.lineNumber ?  this.grammarState[line.lineNumber - 1]?.ruleStack ?? null : null;
+        if(!this.grammar) 
+            return;
+        // Update text content
+        this.documentText[line.lineNumber] = line.text
+        // Don't tokenize line if too long
+        if(line.text.length > 20000) {
+            this.tokensArray[line.lineNumber] = undefined
+            return
+        }
+        // Tokenize line
+        const prevState = line.lineNumber ?  this.tokensArray[line.lineNumber - 1]?.ruleStack ?? null : null;
         const lineTokens = this.grammar.tokenizeLine(line.text, prevState);
-        this.grammarState[line.lineNumber] = lineTokens;
+        this.tokensArray[line.lineNumber] = lineTokens;
     }
     /**
      * @param {vscode.Range} range 
@@ -84,13 +96,13 @@ class DocumentController {
      */
     applyChanges(changes){
         for(let change of changes){
-            const initState = this.grammarState[change.range.end.line].ruleStack
+            // compare ruleStack
+            const initState = this.tokensArray[change.range.end.line]?.ruleStack
             this.reparseRange(this.document.validateRange(change.range))
-            const lastState = this.grammarState[change.range.end.line].ruleStack
+            const lastState = this.tokensArray[change.range.end.line]?.ruleStack
 
             // if (insert line count !== replaced content line count || ruleStack !== init ruleStack) 
             //  {parse the rest of document; return} 
-            
             if( (change.range.end.line - change.range.start.line) !== (change.text.match(/\r\n|\n|\r/g)?.length ?? 0) || initState !== lastState  ){
                 const moreRange = new vscode.Range(change.range.end.line+1, 0 , this.document.lineCount, 0)
                 this.reparseRange(this.document.validateRange(moreRange))
@@ -98,33 +110,30 @@ class DocumentController {
             }
         }
     }
-    /** API: getScopeAt   */
-    /*
-     * @param {vscode.Position} position 
+    /**
+     * API: getScopeAt
+     * @param {vscode.Position} position
      * @returns {TokensInfo}
      */
     getScopeAt(position){
+        // let a = process.hrtime()[1]
         if (!this.grammar)
             return new TokensInfo(new vscode.Range(position,position),"",[]);
         position = this.document.validatePosition(position);
-        // TODO: tokenize line when content changes? 
 
-        if(this.contentChangesArray.length && position.line < this.contentChangesArray[0].range.start.line){
-            const firstChange = this.contentChangesArray[0];
-            const change = [{
-                range : new vscode.Range(position, firstChange.range.end),
-                text : firstChange.text,
-                rangeOffset : firstChange.range.start.compareTo(new vscode.Position(0,0)),
-                rangeLength: firstChange.range.end.compareTo(firstChange.range.start)
-            }]
-            this.applyChanges(change);
+        // FIXME: if some other extensions call this API by changing text without triggering `onDidChangeTextDocument` event in this extension, it may cause error.
+        if(!this.contentChangesArray.length && this.documentText[position.line] !== this.document.lineAt(position.line).text){
+            this.reparseRange(new vscode.Range(position,position))
         }
         this.contentChangesArray.length = 0;
 
-        const token = this.grammarState[position.line]
+        const token = this.tokensArray[position.line]
+        if (!token) 
+            return new TokensInfo(new vscode.Range(position,position),"",[]);
         for (let index = token.tokens.length; index--;) {
             const thisToken = token.tokens[index]
             if (thisToken.startIndex <= position.character ){
+                // console.log(process.hrtime()[1]-a); // TEST SPEED
                 return new TokensInfo(
                     new vscode.Range(position.line, thisToken.startIndex, position.line, thisToken.endIndex),
                     this.document.lineAt(position.line).text.substring(thisToken.startIndex, thisToken.endIndex),
@@ -132,7 +141,7 @@ class DocumentController {
                 )
             }
         }
-    
+
         // default
         return new TokensInfo(new vscode.Range(position,position),"",[]);
     }