+++ CHANGELOG.md
@@ -8,3 +8,15 @@
 
 - Add `refresh()` function for `DocumentController`. 
 - Set a non-empty return value for `getScopeAt()` function.
+
+## [0.0.3]
+
+- FIX: Partly fix a bug that may cause error when call `getScopeAt()` API after changing text.
+
+## [0.0.4]
+
+- NEW: Add "Reload Tokens" Command.
+
+## [0.0.5]
+
+- Other: Several bug fixes. 😝